requirements.txt
opencv-python
pillow
flask
imageio
imageio-ffmpeg